requirements.txt
asgiref==3.5.2
attrs==23.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-cors-headers==3.13.0
django-decouple==2.1
django-jalali==6.0.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
jdatetime==4.1.1
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.1
oauthlib==3.2.1
openapi-codec==1.3.2
packaging==23.2
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0.1
referencing==0.32.1
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.17.1
simplejson==3.19.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
